pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=26.1.0",
    "django>=6.0.3",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.13.0",
    "django-filter>=25.2",
    "django-jazzmin>=3.0.3",
    "django-stubs[compatible-mypy]>=5.2.9",
    "djangorestframework-simplejwt>=5.5.1",
    "gunicorn>=25.3.0",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "psycopg2-binary>=2.9.11",
    "reportlab>=4.4.10",
    "ruff>=0.15.4",
    "whitenoise>=6.12.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "C4", "UP"]

[tool.ruff.lint.per-file-ignores]
"*/migrations/*.py" = ["E501"]

[tool.black]
line-length = 120
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
disallow_untyped_defs = true
warn_return_any = true
warn_unused_configs = true
plugins = ["mypy_django_plugin.main"]
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "config.settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[dependency-groups]
dev = [
    "django-environ>=0.13.0",
    "django-stubs-ext>=5.2.9",
    "djangorestframework-stubs>=3.16.8",
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
]
